Open scenario layers in publication (#350)
* Move download related workers to workers_download module an refactor download logic

* Simplify FileData model

* Use file styling as fallback when publication layer has no styling

* Open publication map layers from scenario as WMS

* Store last http code in network_manager to enable more refined error handling.
