+++ rana_qgis_plugin/utils_data.py
@@ -6,6 +6,14 @@ from typing import Optional
 class DataType(Enum):
     raster = "raster"
     vector = "vector"
+    schematisation = "threedi-schematisation"
+    scenario = "scenario"
+
+    @classmethod
+    def from_value(cls, value: str) -> Optional["DataType"]:
+        if value in cls._value2member_map_:
+            return cls._value2member_map_[value]
+        return None
 
 
 @dataclass
@@ -19,14 +27,15 @@ class RanaPublicationFileData(RanaFileData):
     file_tree: list[str]
     display_name: str  # name for layer in Qgis layer panel
     style_id: Optional[str] = None
+    layer_in_file: Optional[str] = None
 
 
 @dataclass
-class RanaRasterPublicationFileData(RanaPublicationFileData):
-    data_type: DataType = field(default=DataType.raster, init=False)
+class LocalPublicationFileData(RanaPublicationFileData):
+    local_path: str
 
-
-@dataclass
-class RanaVectorPublicationFileData(RanaPublicationFileData):
-    data_type: DataType = field(default=DataType.vector, init=False)
-    layer_in_file: str  # name of layer in gpkg
+    @staticmethod
+    def from_file_data(
+        local_path, file_data: RanaPublicationFileData
+    ) -> "LocalPublicationFileData":
+        return LocalPublicationFileData(local_path=local_path, **file_data.__dict__)
