+++ tests/test_utils.py
@@ -32,25 +32,36 @@ def test_sanitize_path_for_filesystem(input_path, expected_output):
     assert result == expected_output
 
 
+def test_get_local_dir_structure():
+    rana_root = "/root/Rana/"
+    project = "foo"
+    file_id = "baz/bar.txt"
+    file_stem = Path(file_id).stem
+    local_dir = utils.get_local_dir_structure(project, file_id)
+    expected_local_dir = rana_root + project + "/files/baz/" + file_stem
+    assert local_dir == expected_local_dir
+
+
 def test_get_local_file_path():
     rana_root = "/root/Rana/"
     project = "foo"
     file_id = "baz/bar.txt"
     file_name = Path(file_id).name
     file_stem = Path(file_id).stem
-    local_dir, local_path = utils.get_local_file_path(project, file_id)
-    expected_local_dir = rana_root + project + "/files/baz/" + file_stem
-    assert local_dir == expected_local_dir
-    assert local_path == expected_local_dir + "/" + file_name
+    local_path = utils.get_local_file_path(project, file_id)
+    expected_local_path = (
+        rana_root + project + "/files/baz/" + file_stem + "/" + file_name
+    )
+    assert local_path == expected_local_path
 
 
-def test_get_publication_layer_path():
+def test_get_local_publication_dir_structure():
     rana_root = "/root/Rana/"
     project = "foo"
     file_id = "bar.txt"
     file_stem = Path(file_id).stem
     publication_tree = ["publication", "map", "folder"]
-    local_dir, local_path = utils.get_publication_layer_path(
+    local_dir = utils.get_local_publication_dir_structure(
         project, file_id, publication_tree
     )
     publication_tree_path = "/".join(publication_tree)
@@ -58,6 +69,21 @@ def test_get_publication_layer_path():
         rana_root + project + "/publications/" + publication_tree_path + "/" + file_stem
     )
     assert local_dir == expected_local_dir
+
+
+def test_get_local_publication_file_path():
+    rana_root = "/root/Rana/"
+    project = "foo"
+    file_id = "bar.txt"
+    file_stem = Path(file_id).stem
+    publication_tree = ["publication", "map", "folder"]
+    local_path = utils.get_local_publication_file_path(
+        project, file_id, publication_tree
+    )
+    publication_tree_path = "/".join(publication_tree)
+    expected_local_dir = (
+        rana_root + project + "/publications/" + publication_tree_path + "/" + file_stem
+    )
     assert local_path == expected_local_dir + "/" + file_id
